% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abs_path.R
\name{get_abs_path}
\alias{get_abs_path}
\title{Get absolute path}
\usage{
get_abs_path(dir, relpath)
}
\arguments{
\item{dir}{directory}

\item{relpath}{relative path}
}
\description{
Get absolute directory
}
\examples{
  get_abs_path("/Users/htu/myRepo", "scripts")
  # get "/Users/htu/myRepo/scripts"
}
\author{
Hanming Tu
}
